refactor: map filter data와 마커 찍는 로직 연결
